// client/src/app/models/basket.ts
export interface BasketItem {
  productId: number;
  name: string;
  price: number;
  pictureUrl: string;
  brand: string;
  type: string;
  quantity: number;
}

export interface IBasket {
  id: number;
  buyerId: string;
  items: Array<BasketItem>;
  clientSecret?: string;
  paymentIntentId?: string;
}